Changes done as per reviews
